+++ Sources/StreamChat/Database/DTOs/ChannelDTO.swift
@@ -112,11 +112,12 @@ class ChannelDTO: NSManagedObject {
     }
     
     static func loadOrCreate(cid: ChannelId, context: NSManagedObjectContext) -> ChannelDTO {
-        if let existing = Self.load(cid: cid, context: context) {
+        let request = fetchRequest(for: cid)
+        if let existing = load(by: request, context: context).first {
             return existing
         }
         
-        let new = NSEntityDescription.insertNewObject(forEntityName: Self.entityName, into: context) as! ChannelDTO
+        let new = NSEntityDescription.insertNewObject(into: context, for: request)
         new.cid = cid.rawValue
         return new
     }
@@ -167,7 +168,6 @@ extension NSManagedObjectContext {
             )
             dto.extraData = Data()
         }
-        dto.extraData = try JSONEncoder.default.encode(payload.extraData)
         dto.typeRawValue = payload.typeRawValue
         dto.config = payload.config.asDTO(context: self, cid: dto.cid)
         dto.createdAt = payload.createdAt
@@ -401,6 +401,18 @@ extension ChatChannel {
                 .compactMap { try? $0.asModel() }
         }
         
+        let fetchLatestMessageFromUser: () -> ChatMessage? = {
+            guard let currentUser = context.currentUser else { return nil }
+            
+            return try? MessageDTO
+                .loadLastMessage(
+                    from: currentUser.user.id,
+                    in: dto.cid,
+                    context: context
+                )?
+                .asModel()
+        }
+        
         let fetchWatchers: () -> [ChatUser] = {
             UserDTO
                 .loadLastActiveWatchers(cid: cid, context: context)
@@ -448,6 +460,7 @@ extension ChatChannel {
             extraData: extraData,
             //            invitedMembers: [],
             latestMessages: { fetchMessages() },
+            lastMessageFromCurrentUser: { fetchLatestMessageFromUser() },
             pinnedMessages: { dto.pinnedMessages.compactMap { try? $0.asModel() } },
             muteDetails: fetchMuteDetails,
             previewMessage: { try? dto.previewMessage?.asModel() },