+++ Sources/StreamChat/Database/DTOs/MessageDTO.swift
@@ -164,6 +164,14 @@ class MessageDTO: NSManagedObject {
     private static func nonDeletedMessagesPredicate() -> NSPredicate {
         .init(format: "deletedAt == nil")
     }
+    
+    private static func channelPredicate(with cid: String) -> NSPredicate {
+        .init(format: "channel.cid == %@", cid)
+    }
+    
+    private static func messageSentPredicate() -> NSPredicate {
+        .init(format: "localMessageStateRaw == nil")
+    }
 
     /// Returns predicate for displaying messages after the channel truncation date.
     private static func nonTruncatedMessagesPredicate() -> NSCompoundPredicate {
@@ -192,10 +200,6 @@ class MessageDTO: NSManagedObject {
         deletedMessagesVisibility: ChatClientConfig.DeletedMessageVisibility,
         shouldShowShadowedMessages: Bool
     ) -> NSCompoundPredicate {
-        let channelMessage = NSPredicate(
-            format: "channel.cid == %@", cid
-        )
-
         let channelMessagePredicate = NSCompoundPredicate(orPredicateWithSubpredicates: [
             .init(format: "showReplyInChannel == 1"),
             .init(format: "parentMessageId == nil")
@@ -219,7 +223,7 @@ class MessageDTO: NSManagedObject {
         ])
         
         var subpredicates = [
-            channelMessage,
+            channelPredicate(with: cid),
             channelMessagePredicate,
             messageTypePredicate,
             nonTruncatedMessagesPredicate(),
@@ -341,9 +345,7 @@ class MessageDTO: NSManagedObject {
     }
     
     static func load(id: String, context: NSManagedObjectContext) -> MessageDTO? {
-        let request = NSFetchRequest<MessageDTO>(entityName: entityName)
-        request.predicate = NSPredicate(format: "id == %@", id)
-        return load(by: request, context: context).first
+        load(by: id, context: context).first
     }
     
     static func loadOrCreate(id: String, context: NSManagedObjectContext) -> MessageDTO {
@@ -351,7 +353,8 @@ class MessageDTO: NSManagedObject {
             return existing
         }
         
-        let new = NSEntityDescription.insertNewObject(forEntityName: entityName, into: context) as! Self
+        let request = fetchRequest(id: id)
+        let new = NSEntityDescription.insertNewObject(into: context, for: request)
         new.id = id
         new.latestReactions = []
         new.ownReactions = []
@@ -380,11 +383,11 @@ class MessageDTO: NSManagedObject {
         context: NSManagedObjectContext
     ) -> [MessageDTO] {
         let subpredicates: [NSPredicate] = [
-            .init(format: "channel.cid == %@", cid),
+            channelPredicate(with: cid),
             .init(format: "user.currentUser != nil"),
             .init(format: "createdAt > %@", createdAtFrom as NSDate),
             .init(format: "createdAt <= %@", createdAtThrough as NSDate),
-            .init(format: "localMessageStateRaw == nil"),
+            messageSentPredicate(),
             nonTruncatedMessagesPredicate(),
             nonDeletedMessagesPredicate()
         ]
@@ -400,6 +403,19 @@ class MessageDTO: NSManagedObject {
         let request = NSFetchRequest<ChannelReadDTO>(entityName: ChannelReadDTO.entityName)
         request.predicate = NSPredicate(format: "readMessagesFromCurrentUser.id CONTAINS %@", messageId)
         return (try? context.count(for: request)) ?? 0
+    }
+    
+    static func loadLastMessage(from userId: String, in cid: String, context: NSManagedObjectContext) -> MessageDTO? {
+        let request = NSFetchRequest<MessageDTO>(entityName: entityName)
+        request.predicate = NSCompoundPredicate(andPredicateWithSubpredicates: [
+            channelPredicate(with: cid),
+            .init(format: "user.id == %@", userId),
+            .init(format: "type != %@", MessageType.ephemeral.rawValue),
+            messageSentPredicate()
+        ])
+        request.sortDescriptors = [NSSortDescriptor(keyPath: \MessageDTO.createdAt, ascending: false)]
+        request.fetchLimit = 1
+        return load(by: request, context: context).first
     }
 }