+++ StreamChatUITestsAppUITests/Robots/ParticipantRobot.swift
@@ -59,10 +59,11 @@ final class ParticipantRobot: Robot {
         return self
     }
     
-    // Think twice before using it
+    // Sleep in seconds
     @discardableResult
-    func chill(duration: UInt32) -> Self {
-        sleep(duration)
+    func wait(_ duration: TimeInterval) -> Self {
+        let sleepTime = UInt32(duration * 1000)
+        usleep(sleepTime)
         return self
     }
     
@@ -78,6 +79,7 @@ final class ParticipantRobot: Robot {
     
     @discardableResult
     func sendMessage(_ text: String) -> Self {
+        waitForChannelQueryUpdate()
         server.websocketMessage(
             text,
             channelId: server.currentChannelId,
@@ -85,6 +87,21 @@ final class ParticipantRobot: Robot {
             eventType: .messageNew,
             user: participant()
         )
+        return self
+    }
+
+    /// The given text will be decorated with the index, eg "message-10"
+    @discardableResult
+    func sendMultipleMessages(repeatingText text: String, count: Int) -> Self {
+        var texts = [String]()
+        for index in 1...count {
+            texts.append("\(text)-\(index)")
+        }
+
+        texts.forEach {
+            sendMessage($0)
+            wait(0.5)
+        }
         return self
     }
     
@@ -103,12 +120,17 @@ final class ParticipantRobot: Robot {
     
     @discardableResult
     func deleteMessage() -> Self {
-        let messageId = server.lastMessage?[MessagePayloadsCodingKeys.id.rawValue] as? String
+        let user = participant()
+        guard let userId = user?[UserPayloadsCodingKeys.id.rawValue] as? String else {
+            return self
+        }
+        let message = server.findMessageByUserId(userId)
+        let messageId = message?[MessagePayloadsCodingKeys.id.rawValue] as? String
         server.websocketMessage(
             channelId: server.currentChannelId,
             messageId: messageId,
             eventType: .messageDeleted,
-            user: participant()
+            user: user
         )
         return self
     }
@@ -180,5 +202,12 @@ final class ParticipantRobot: Robot {
         }
 
         return server.setUpUser(source: message, details: _user)
+    }
+    
+    // To avoid collision of the first channel query call and a first participant's message
+    private func waitForChannelQueryUpdate(timeout: Double = 5) {
+        let endTime = Date().timeIntervalSince1970 * 1000 + timeout * 1000
+        while !server.channelQueryEndpointWasCalled
+                && endTime > Date().timeIntervalSince1970 * 1000 {}
     }
 }