+++ StreamChatUITestsAppUITests/Tests/StreamTestCase.swift
@@ -45,8 +45,9 @@ extension StreamTestCase {
     
     enum Tags: String {
         case coreFeatures = "Core Features"
+        case slowMode = "Slow Mode"
         case offlineSupport = "Offline Support"
-        case messageReceipts = "Message Receipts"
+        case messageDeliveryStatus = "Message Delivery Status"
     }
     
     func addTags(_ tags: [Tags]) {