+++ Tests/StreamChatTests/Database/DTOs/UserDTO_Tests.swift
@@ -108,25 +108,21 @@ final class UserDTO_Tests: XCTestCase {
         }
     }
     
-    func test_DTO_asPayload() {
+    func test_DTO_asPayload() throws {
         let userId = UUID().uuidString
         
         let payload: UserPayload = .dummy(userId: userId, extraData: ["k": .string("v")])
         
         // Asynchronously save the payload to the db
-        database.write { session in
+        try database.writeSynchronously { session in
             try! session.saveUser(payload: payload)
         }
         
         // Load the user from the db and check the fields are correct
-        var loadedUserPayload: UserRequestBody? {
-            database.viewContext.user(id: userId)?.asRequestBody()
-        }
+        let loadedUserPayload = database.viewContext.user(id: userId)?.asRequestBody()
         
-        AssertAsync {
-            Assert.willBeEqual(payload.id, loadedUserPayload?.id)
-            Assert.willBeEqual(payload.extraData, loadedUserPayload?.extraData)
-        }
+        XCTAssertEqual(payload.id, loadedUserPayload?.id)
+        XCTAssertEqual(payload.extraData, loadedUserPayload?.extraData)
     }
     
     func test_DTO_resetsItsEphemeralValues() throws {