+++ Tests/StreamChatUITests/SnapshotTests/Composer/ComposerVC_Tests.swift
@@ -261,4 +261,19 @@ final class ComposerVC_Tests: XCTestCase {
         
         AssertSnapshot(containerVC, variants: [.defaultLight])
     }
+    
+    func test_channelWithSlowModeActive_messageIsSent_SlowModeIsOnWithCountdownShown() {
+        // GIVEN
+        let composerVC = ComposerVC()
+        composerVC.cooldownTracker = CooldownTracker_Mock(timer: ScheduledStreamTimer_Mock())
+        composerVC.appearance = Appearance.default
+        composerVC.content.text = "Test text"
+        composerVC.viewDidLoad()
+        
+        // WHEN
+        composerVC.cooldownTracker.start(with: 120)
+        
+        // THEN
+        AssertSnapshot(composerVC)
+    }
 }